// fuelmap-app/src/types/enums.ts
export enum TaskType {
  SECTOR = 0,
  SYSTEM = 1,
  PLANET = 2,
  MOON = 3,
}

export enum TaskColor {
  VIOLET = "#9488F0",
  GREEN = "#6FDCBA",
  BLUE = "#3CA1ED",
  RED = "#F33F5B",
  ORANGE = "#F98654",
  YELLOW = "#F6D083",
  BROWN = "#BC768B",
}

export const COLORS = Object.values(TaskColor);